# src/game_engine/static/openapi.yml
openapi: 3.0.3
info:
  title: Card Game API
  description: API per il Game Engine del gioco di carte.
  version: 1.0.1
servers:
  - url: /game
    description: Percorso base del Game Engine (Blueprint prefix)

# =============================================================
#  DEFINIZIONE DEI PERCORSI (ENDPOINTS)
# =============================================================
paths:
  /connect:
    post:
      summary: Connette un giocatore al matchmaking
      description: Aggiunge un giocatore (tramite UUID e nome) alla lista 'online_players' in attesa di una partita.
      tags:
        - Matchmaking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Il nome utente del giocatore.
                  example: "Alice"
                uuid:
                  type: string
                  description: L'UUID univoco del giocatore.
                  example: "user-alice-uuid"
              required:
                - username
                - uuid
      responses:
        '200':
          description: Giocatore connesso o già online.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Alice connected."
                  players_online_count:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'

  /matchmake:
    post:
      summary: Cerca una partita
      description: Se ci sono almeno 2 giocatori online, li accoppia, crea una partita e li rimuove dalla coda.
      tags:
        - Matchmaking
      responses:
        '200':
          description: Risultato del matchmaking.
          content:
            application/json:
              schema:
                oneOf: # Può restituire uno di questi due oggetti
                  - $ref: '#/components/schemas/WaitingForMatch'
                  - $ref: '#/components/schemas/MatchCreated'
  
  /start:
    post:
      summary: Avvia una nuova partita (Manuale/Test)
      description: Crea manualmente una nuova istanza di gioco per due giocatori specifici.
      tags:
        - Game Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player1_uuid:
                  type: string
                  example: "user-alice-uuid"
                player1_name:
                  type: string
                  example: "Alice"
                player2_uuid:
                  type: string
                  example: "user-bob-uuid"
                player2_name:
                  type: string
                  example: "Bob"
              required:
                - player1_uuid
                - player1_name
                - player2_uuid
                - player2_name
      responses:
        '201':
          description: Partita creata con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  game_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: "started"
        '400':
          $ref: '#/components/responses/BadRequest'

  /deck/{game_id}:
    post:
      summary: Seleziona lo slot del mazzo
      description: Un giocatore invia il suo UUID e lo slot del mazzo scelto (1-5). Il server contatterà il 'collection-service' per recuperare e validare il mazzo.
      tags:
        - Game Flow
      parameters:
        - name: game_id
          in: path
          required: true
          description: L'ID della partita.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_uuid:
                  type: string
                  description: L'UUID del giocatore che sceglie il mazzo.
                  example: "user-alice-uuid"
                deck_slot:
                  type: integer
                  description: Lo slot del mazzo (es. 1, 2, 3...).
                  example: 1
              required:
                - player_uuid
                - deck_slot
      responses:
        '200':
          description: Mazzo selezionato. La risposta indica se si attende l'avversario o se la partita è iniziata.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "Alice deck selected. Waiting for opponent."
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "Alice deck selected. Both ready! Game started, 3 cards drawn."
        '400':
          $ref: '#/components/responses/BadRequest' # Es. "player_uuid and deck_slot are required", "Deck not found", "Suit exceeds 15 points"
        '500':
          $ref: '#/components/responses/ServerError' # Es. "Could not reach collection service"

  /play/{game_id}:
    post:
      summary: Gioca una carta
      description: Un giocatore (identificato da UUID) gioca una carta dalla sua mano.
      tags:
        - Game Flow
      parameters:
        - name: game_id
          in: path
          required: true
          description: L'ID della partita.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                player_uuid:
                  type: string
                  description: L'UUID del giocatore che gioca la carta.
                  example: "user-alice-uuid"
                card:
                  $ref: '#/components/schemas/Card'
              required:
                - player_uuid
                - card
      responses:
        '200':
          description: Mossa eseguita. La risposta indica lo stato (in attesa o risolto).
          content:
            application/json:
              schema:
                oneOf: # La risposta sarà uno di questi due schemi
                  - $ref: '#/components/schemas/WaitingResponse'
                  - $ref: '#/components/schemas/ResolvedResponse'
        '400':
          $ref: '#/components/responses/BadRequest' # Es. "card is required"
        '404':
          $ref: '#/components/responses/NotFound' # Es. "Invalid game ID", "card not in hand"

  /state/{game_id}:
    get:
      summary: Recupera lo stato della partita
      description: Ottiene lo stato attuale del gioco, inclusi punteggi, turni e vincitore.
      tags:
        - Game Management
      parameters:
        - name: game_id
          in: path
          required: true
          description: L'ID della partita.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stato attuale della partita.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '404':
          $ref: '#/components/responses/NotFound' # Es. "Invalid game ID"

# =============================================================
#  COMPONENTI RIUTILIZZABILI
# =============================================================
components:
  schemas:
    Card:
      type: object
      properties:
        value:
          type: string
          example: "K"
        suit:
          type: string
          example: "clubs"
      required:
        - value
        - suit

    PlayerState:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        hand_size:
          type: integer
          example: 3
        # Aggiungi 'hand' se vuoi esporlo (come abbiamo fatto per il test)
        # hand:
        #   type: array
        #   items:
        #     $ref: '#/components/schemas/Card'
            
    GameState:
      description: Lo stato completo di una partita.
      type: object
      properties:
        game_id:
          type: string
          format: uuid
        turn_number:
          type: integer
          example: 3
        winner:
          type: string
          nullable: true
          example: null
        scores:
          type: object
          additionalProperties:
            type: integer
          example: {"Alice": 2, "Bob": 1}
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerState'
        turn_history:
          type: array
          description: Storico dei turni giocati.
          items:
            type: object # Questo schema può essere dettagliato ulteriormente se necessario
            
    WaitingResponse:
      description: La risposta inviata quando il primo giocatore ha giocato e si attende il secondo.
      type: object
      properties:
        status:
          type: string
          example: "waiting"

    ResolvedResponse:
      description: La risposta inviata dopo che entrambi hanno giocato e il round è stato risolto.
      type: object
      properties:
        status:
          type: string
          example: "resolved"
        message:
          type: string
          example: "Alice wins round 1!"
        scores:
          type: object
          additionalProperties:
            type: integer
          example: {"Alice": 1, "Bob": 0}
        turn_number:
          type: integer
          example: 1
        match_winner:
          type: string
          nullable: true
          example: null

    WaitingForMatch:
      type: object
      properties:
        message:
          type: string
          example: "Waiting for more players..."
        players_online_count:
          type: integer
          example: 1
          
    MatchCreated:
      type: object
      properties:
        message:
          type: string
          example: "Match created between Alice and Bob"
        game_id:
          type: string
          format: uuid
        players:
          type: array
          items:
            type: object
            properties:
              uuid:
                type: string
                format: uuid
              name:
                type: string
            example: [{"uuid": "user-alice-uuid", "name": "Alice"}, {"uuid": "user-bob-uuid", "name": "Bob"}]

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

  responses:
    BadRequest:
      description: Richiesta non valida (es. dati mancanti, validazione fallita).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example: {"error": "player_uuid and deck_slot are required"}
    NotFound:
      description: Risorsa non trovata (es. game_id non valido).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example: {"error": "Invalid game ID"}
    ServerError:
      description: Errore interno del server (es. impossibile contattare un altro servizio).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example: {"error": "Could not reach collection service"}